requirements/development.txt
-r base.txt # includes the base.txt requirements file

django-debug-toolbar==1.4
django-extensions==1.6.1
